Move _make_mask to utils
